Move python worker tasks into tasks folder
